refactor(constructor): Refactor constructor widgets on client
- Made constructor entity's named widget folder
- Removed constructor list to constructor entity's named widget ui segment
