+++ packages/client/src/widgets/constructor/ui/index.ts
@@ -1,0 +1,1 @@
+export { ConstructorList } from './list';
